community-mind-mirror/requirements.txt
# Async HTTP
httpx>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# RSS/XML parsing
feedparser>=6.0.0
lxml>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Logging
structlog>=24.0.0

# Retries
tenacity>=8.0.0

# Redis
redis>=5.0.0

# Web framework (for future API)
fastapi>=0.110.0
uvicorn>=0.27.0

# YouTube
youtube-transcript-api>=1.0.0
google-api-python-client>=2.100.0

# Job market
python-jobspy>=1.1.0

# Scheduler
apscheduler>=3.10.0

# Utilities
beautifulsoup4>=4.12.0

# LLM (Azure OpenAI)
openai>=1.30.0

# Sentiment analysis
vaderSentiment>=3.3.2

# Graph analysis
networkx>=3.2
python-louvain>=0.16
